requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
bcrypt==4.0.1
PyJWT==2.8.0
gunicorn==21.2.0
pymongo==4.5.0
